[add] framebuffer Class and EditorViewport class need to link them

+++ undefined/source/src/resources/model.cpp
@@ -47,8 +47,8 @@ void Model::Draw()
 
     for (std::pair<std::shared_ptr<Mesh>, std::shared_ptr<Texture>> pair : mModel)
     {
-        Render->SetBufferData(GL_ARRAY_BUFFER, pair.first->Vertices.size() * sizeof(Vertex), &pair.first->Vertices[0], GL_STATIC_DRAW);
-        Render->SetBufferData(GL_ELEMENT_ARRAY_BUFFER, pair.first->Indices.size() * sizeof(unsigned int), &pair.first->Indices[0], GL_STATIC_DRAW);
+        Render->SetBufferData(GL_ARRAY_BUFFER, (int)pair.first->Vertices.size() * sizeof(Vertex), &pair.first->Vertices[0], GL_STATIC_DRAW);
+        Render->SetBufferData(GL_ELEMENT_ARRAY_BUFFER, (int)pair.first->Indices.size() * sizeof(unsigned int), &pair.first->Indices[0], GL_STATIC_DRAW);
 
         if (pair.second)
         {
